pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "First version of project."
authors = ["Mikhail Uvarov <celovekvodichka@gmail.com>"]
readme = "README.md"
repository = "https://github.com/Celovechek/python-project-50/"
classifiers = [
        "Natural Language :: English",
        "Operating System :: POSIX :: Linux",
        "Programming Language :: Python :: 3.6",
        "Programming Language :: Python :: 3.7",
        "Programming Language :: Python :: 3.8",
        "Programming Language :: Python :: 3.9",
        "Programming Language :: Python :: 3.10",
        "Programming Language :: Python :: 3.11",
        "Programming Language :: Python :: 3.12",
        "Programming Language :: Python :: 3.13",
        "Topic :: Education",
]
packages = [
        { include = "gendiff" },
]


[tool.poetry.dependencies]
python = "^3.10"
prompt = "^0.4.1"
pytest-cov = "^5.0.0"
pyyaml = "^6.0.2"


[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
